move account component files
